pyproject.toml
[project]
name = "ai-landscape"
version = "2025.1"
description = "Living AI ecosystem landscape diagram — grows as AI grows"
requires-python = ">=3.12"
dependencies = [
    "diagrams>=0.25.1",
    "pillow>=12.1.1",
    "reportlab>=4.4.10",
    "requests>=2.33.0",
]
